LayoutTests/storage/indexeddb/intversion-close-between-events-expected.txt

Try to call db.close() after upgradeneeded but before the corresponding success event fires

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-close-between-events.html"
indexedDB.deleteDatabase(dbname)
request = indexedDB.open(dbname, 7)


openSuccess():
PASS didCallCloseDB is false
FAIL sawTransactionComplete should be true. Was false.
db = event.target.result
FAIL db.version should be 7 (of type number). Was  (of type string).
PASS successfullyParsed is true

TEST COMPLETE
